feat: 본인이 쓴 댓글이 아닌데 삭제를 시도하면 예외 처리 기능 구현

+++ backend/src/main/java/com/staccato/comment/service/CommentService.java
@@ -74,6 +74,9 @@ public class CommentService {
 
     @Transactional
     public void deleteComment(long commentId, Member member) {
-        commentRepository.deleteById(commentId);
+        commentRepository.findById(commentId).ifPresent(comment -> {
+            validateCommentOwner(comment, member);
+            commentRepository.deleteById(commentId);
+        });
     }
 }
